requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
numpy==1.26.2
scikit-learn==1.3.2
channels==4.0.0
daphne>=4.0.0
